Moved lots of date checks to JS
Rather than setting date checks in forms.py, date checks are now set 
using JS in the html pages. This means that "today" will always be 
today, not "today" when the server restarted (not an issue on 
PythonAnywhere as it's constantly re-started, but would be an issue if 
moved to a dedicated server).

Also set it so that when booking in a new item, the expiry date 
dynamically changes so that the minimum value is the date received

This is part of issue #33