Able to change location of an item using "edit inventory item"
On that screen the location is a hyperlink that takes you to a page to change the location of an item.
You can't change it to the same location.
This closes #39

Also tidied up the "edit inventory item" page as some items weren't lining up correctly
Also removed some "batch IDs" that should have been "Stock Numbers"
Also updated the database version number